Create repository layer, replace int type id to Long type

+++ src/main/java/com/example/instazoo/entity/Post.java
@@ -18,7 +18,7 @@ import java.util.Set;
 public class Post {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String title;
     private String description;
     private String location;
